Upgrade npm packages and remove hompage/PUBLIC_URL
Running `npm install` caused warning about security. I ran the following:

    npm audit fix

    npm aduit fix --force

    npm upgrade

Also removed `homepage` field from package.json. This is supposed to be the
relative point to the app, not the source code's project home page. Removing
this also means we can remove the PUBLIC_URL env var in the build script
command which was causing problems when building on Windows.