Add Option to Save Attachments to Disk
Added the options:

--save       Flag to save to enable files to disk
--savepath   Dir to save files to relative to script (default: ./downloads/)

Was running into a problem where I wanted to test many emails with
attachments. The current method of rendering the page sends all attachments
embedded in the JSON. This causes very long page loads for even a few emails.

I decided to add an option to instead save the attachments to disk when the
email is parsed. The code still uses the simpleParse method so initially
attachments are buffered into memory. However, after it is parsed and the
option is enabled it is saved to disk and the `attachment.content` property
is deleted and should be released from memory.

Since we are saving to disk a new API was needed. Each attachment saved to
disk gets a `shortid` to make it unique. This ID is also sent in the API call
and using express's res.sendFile() makes it a one liner.

The button openAttachment method was also updated in App.js so that if an
attachment was saved then it will use the API as a link.

To help with clean up the delete emails API will also delete any attachment
files it knows of. It doesn't delete everything in the save folder as it
might accidentally delete files that the user didn't want to be deleted.
Manual clean up will be needed if the service is restarted.

This has helped improve load time of the web page for display of the emails.